+++ dev/com.ibm.ws.wsat.recovery_fat.multi.2/fat/src/tests/MultiRecoveryTest2.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2023 IBM Corporation and others.
+ * Copyright (c) 2020, 2024 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -19,7 +19,7 @@ import componenttest.annotation.AllowedFFDC;
 import componenttest.annotation.ExpectedFFDC;
 import componenttest.custom.junit.runner.FATRunner;
 
-@AllowedFFDC(value = { "com.ibm.tx.jta.ut.util.AlreadyDumpedException", "javax.transaction.SystemException", "javax.transaction.xa.XAException", "java.io.IOException", "java.io.EOFException", "org.apache.cxf.binding.soap.SoapFault", "javax.xml.stream.XMLStreamException" })
+@AllowedFFDC(value = { "com.ibm.tx.jta.ut.util.AlreadyDumpedException", "javax.transaction.SystemException", "javax.transaction.xa.XAException", "java.io.IOException", "java.io.EOFException", "org.apache.cxf.binding.soap.SoapFault", "javax.xml.stream.XMLStreamException", "com.ctc.wstx.exc.WstxIOException" })
 @RunWith(FATRunner.class)
 public class MultiRecoveryTest2 extends MultiRecoveryTest {
 
@@ -49,67 +49,26 @@ public class MultiRecoveryTest2 extends MultiRecoveryTest {
 	public void WSTXMPR006AFVT() throws Exception {
 		recoveryTest(server1, server2, "601","server1");
 	}
-	
+
+	@Test
+	public void WSTXMPR003CFVT() throws Exception {
+		recoveryTest(server1, server2, "303","both");
+	}
+
+	@Test
+	@AllowedFFDC(value = {"javax.xml.ws.WebServiceException" })
+	public void WSTXMPR004AFVT() throws Exception {
+		recoveryTest(server1, server2, "401","server1");
+	}
+
 	@Test
 	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
-	public void WSTXMPR006BFVT() throws Exception {
-		recoveryTest(server1, server2, "602","server2");
+	public void WSTXMPR004BFVT() throws Exception {
+		recoveryTest(server1, server2, "402","server2");
 	}
-	
+
 	@Test
-	public void WSTXMPR006CFVT() throws Exception {
-		recoveryTest(server1, server2, "603","both");
-	}
-	
-	@Test
-	public void WSTXMPR007AFVT() throws Exception {
-		recoveryTest(server1, server2, "701","server1");
-	}
-	
-	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException" })
-	@AllowedFFDC(value = { "java.util.concurrent.RejectedExecutionException", "com.ibm.ws.Transaction.JTA.HeuristicHazardException" })
-	// JDK8: Allow HeuristicHazardException
-	public void WSTXMPR007BFVT() throws Exception {
-		recoveryTest(server1, server2, "702","server2");
-	}
-	
-	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException" })
-	@AllowedFFDC(value = { "com.ibm.ws.Transaction.JTA.HeuristicHazardException" })
-	// Need Jon Review:
-	// javax.transaction.xa.XAException 
-	// Caused by: com.ibm.tx.jta.XAResourceNotAvailableException
-	// Need review on whether it is expected
-	// Report javax.transaction.SystemException
-	// JDK8: Allow HeuristicHazardException
-	public void WSTXMPR007CFVT() throws Exception {
-		recoveryTest(server1, server2, "703","both");
-	}
-	
-	@Test
-	public void WSTXMPR008AFVT() throws Exception {
-		recoveryTest(server1, server2, "801","server1");
-	}
-	
-	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException" })
-	@AllowedFFDC(value = { "com.ibm.ws.Transaction.JTA.HeuristicHazardException" })
-	// JDK8: Allow HeuristicHazardException
-	public void WSTXMPR008BFVT() throws Exception {
-		recoveryTest(server1, server2, "802","server2");
-	}
-	
-	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException" })
-	@AllowedFFDC(value = { "com.ibm.ws.Transaction.JTA.HeuristicHazardException" })
-	// Need Jon Review:
-	// javax.transaction.xa.XAException 
-	// Caused by: com.ibm.tx.jta.XAResourceNotAvailableException
-	// Need review on whether it is expected
-	// Report javax.transaction.SystemException
-	// JDK8: Allow HeuristicHazardException
-	public void WSTXMPR008CFVT() throws Exception {
-		recoveryTest(server1, server2, "803","both");
+	public void WSTXMPR004CFVT() throws Exception {
+		recoveryTest(server1, server2, "403","both");
 	}
 }
