+++ dev/com.ibm.ws.wsat.recovery_fat.multi.3/fat/src/tests/MultiRecoveryTest3.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2023 IBM Corporation and others.
+ * Copyright (c) 2020, 2024 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -24,83 +24,65 @@ import componenttest.custom.junit.runner.FATRunner;
 public class MultiRecoveryTest3 extends MultiRecoveryTest {
 
 	@Test
-	@AllowedFFDC(value = {"javax.xml.ws.WebServiceException"})
-	public void WSTXMPR009AFVT() throws Exception {
-		recoveryTest(server1, server2, "901","server1");
+	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
+	public void WSTXMPR006BFVT() throws Exception {
+		recoveryTest(server1, server2, "602","server2");
 	}
-	
+
 	@Test
-	// XAException should be expected here but halt is no longer halting the JVM so extra protocol
-	// messages can get through after the dumpState.
-	public void WSTXMPR009BFVT() throws Exception {
-		recoveryTest(server1, server2, "902","server2");
+	public void WSTXMPR006CFVT() throws Exception {
+		recoveryTest(server1, server2, "603","both");
 	}
-	
+
+	@Test
+	public void WSTXMPR007AFVT() throws Exception {
+		recoveryTest(server1, server2, "701","server1");
+	}
+
 	@Test
 	@ExpectedFFDC(value = {"javax.transaction.xa.XAException" })
-	//javax.transaction.xa.XAException 
-	//Caused by: com.ibm.tx.jta.XAResourceNotAvailableException
-	//Need review on whether it is expected
-	public void WSTXMPR009CFVT() throws Exception {
-		recoveryTest(server1, server2, "903","both");
+	@AllowedFFDC(value = { "java.util.concurrent.RejectedExecutionException", "com.ibm.ws.Transaction.JTA.HeuristicHazardException" })
+	// JDK8: Allow HeuristicHazardException
+	public void WSTXMPR007BFVT() throws Exception {
+		recoveryTest(server1, server2, "702","server2");
 	}
-	
-	@Test
-	@AllowedFFDC(value = {"javax.xml.ws.WebServiceException"})
-	public void WSTXMPR010AFVT() throws Exception {
-		recoveryTest(server1, server2, "1001","server1");
-	}
-	
-	@Test
-	@AllowedFFDC(value = {"javax.xml.ws.WebServiceException"})
-	// XAException should be expected here but halt is no longer halting the JVM so extra protocol
-	// messages can get through after the dumpState.
-	public void WSTXMPR010BFVT() throws Exception {
-		recoveryTest(server1, server2, "1002","server2");
-	}
-	
+
 	@Test
 	@ExpectedFFDC(value = {"javax.transaction.xa.XAException" })
-	// Can get HeuristicRollbackException if commit gets through to subordinate after resources were already dumped
-	@AllowedFFDC(value={"javax.transaction.HeuristicRollbackException"})
-	public void WSTXMPR010CFVT() throws Exception {
-		recoveryTest(server1, server2, "1003","both");
+	@AllowedFFDC(value = { "com.ibm.ws.Transaction.JTA.HeuristicHazardException" })
+	// Need Jon Review:
+	// javax.transaction.xa.XAException 
+	// Caused by: com.ibm.tx.jta.XAResourceNotAvailableException
+	// Need review on whether it is expected
+	// Report javax.transaction.SystemException
+	// JDK8: Allow HeuristicHazardException
+	public void WSTXMPR007CFVT() throws Exception {
+		recoveryTest(server1, server2, "703","both");
 	}
-	
+
 	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
-	public void WSTXMPR011AFVT() throws Exception {
-		recoveryTest(server1, server2, "1101","server1");
+	public void WSTXMPR008AFVT() throws Exception {
+		recoveryTest(server1, server2, "801","server1");
 	}
-	
+
 	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
-	public void WSTXMPR011BFVT() throws Exception {
-		recoveryTest(server1, server2, "1102","server2");
+	@ExpectedFFDC(value = {"javax.transaction.xa.XAException" })
+	@AllowedFFDC(value = { "com.ibm.ws.Transaction.JTA.HeuristicHazardException" })
+	// JDK8: Allow HeuristicHazardException
+	public void WSTXMPR008BFVT() throws Exception {
+		recoveryTest(server1, server2, "802","server2");
 	}
-	
+
 	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
-	public void WSTXMPR011CFVT() throws Exception {
-		recoveryTest(server1, server2, "1103","both");
-	}
-	
-	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
-	@AllowedFFDC(value = {"javax.xml.ws.WebServiceException", "com.ibm.ws.wsat.service.WSATException" })
-	public void WSTXMPR012AFVT() throws Exception {
-		recoveryTest(server1, server2, "1201","server1");
-	}
-	
-	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
-	public void WSTXMPR012BFVT() throws Exception {
-		recoveryTest(server1, server2, "1202","server2");
-	}
-	
-	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
-	public void WSTXMPR012CFVT() throws Exception {
-		recoveryTest(server1, server2, "1203","both");
+	@ExpectedFFDC(value = {"javax.transaction.xa.XAException" })
+	@AllowedFFDC(value = { "com.ibm.ws.Transaction.JTA.HeuristicHazardException" })
+	// Need Jon Review:
+	// javax.transaction.xa.XAException 
+	// Caused by: com.ibm.tx.jta.XAResourceNotAvailableException
+	// Need review on whether it is expected
+	// Report javax.transaction.SystemException
+	// JDK8: Allow HeuristicHazardException
+	public void WSTXMPR008CFVT() throws Exception {
+		recoveryTest(server1, server2, "803","both");
 	}
 }
