+++ dev/com.ibm.ws.wsat.recovery_fat.multi.4/fat/src/tests/MultiRecoveryTest4.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2023 IBM Corporation and others.
+ * Copyright (c) 2020, 2024 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -20,76 +20,63 @@ import componenttest.annotation.ExpectedFFDC;
 import componenttest.custom.junit.runner.FATRunner;
 
 @SuppressWarnings("restriction")
-@AllowedFFDC(value = { "com.ibm.tx.jta.ut.util.AlreadyDumpedException", "javax.transaction.SystemException", "javax.transaction.xa.XAException", "java.io.IOException", "java.io.EOFException", "org.apache.cxf.binding.soap.SoapFault", "javax.xml.stream.XMLStreamException" })
+@AllowedFFDC(value = { "com.ibm.tx.jta.ut.util.AlreadyDumpedException", "javax.transaction.SystemException", "javax.transaction.xa.XAException", "java.io.IOException", "java.io.EOFException", "org.apache.cxf.binding.soap.SoapFault", "javax.xml.stream.XMLStreamException", "com.ctc.wstx.exc.WstxIOException" })
 @RunWith(FATRunner.class)
 public class MultiRecoveryTest4 extends MultiRecoveryTest {
 
 	@Test
-	@AllowedFFDC(value = {"javax.xml.ws.WebServiceException", "com.ibm.ws.wsat.service.WSATException" })
-	public void WSTXMPR013AFVT() throws Exception {
-		recoveryTest(server1, server2, "1301","server1");
+	@AllowedFFDC(value = {"javax.xml.ws.WebServiceException"})
+	public void WSTXMPR009AFVT() throws Exception {
+		recoveryTest(server1, server2, "901","server1");
 	}
-	
+
 	@Test
-	@AllowedFFDC(value = { "javax.xml.ws.WebServiceException", "com.ibm.ws.wsat.service.WSATException" })
-	@ExpectedFFDC(value = {"javax.transaction.RollbackException"})
-	public void WSTXMPR013BFVT() throws Exception {
-		recoveryTest(server1, server2, "1302","server2");
+	// XAException should be expected here but halt is no longer halting the JVM so extra protocol
+	// messages can get through after the dumpState.
+	public void WSTXMPR009BFVT() throws Exception {
+		recoveryTest(server1, server2, "902","server2");
 	}
-	
+
 	@Test
-	public void WSTXMPR013CFVT() throws Exception {
-		recoveryTest(server1, server2, "1303","both");
+	@ExpectedFFDC(value = {"javax.transaction.xa.XAException" })
+	//javax.transaction.xa.XAException 
+	//Caused by: com.ibm.tx.jta.XAResourceNotAvailableException
+	//Need review on whether it is expected
+	public void WSTXMPR009CFVT() throws Exception {
+		recoveryTest(server1, server2, "903","both");
 	}
-	
+
 	@Test
-	public void WSTXMPR014AFVT() throws Exception {
-		recoveryTest(server1, server2, "1401","server1");
+	@AllowedFFDC(value = {"javax.xml.ws.WebServiceException"})
+	public void WSTXMPR010AFVT() throws Exception {
+		recoveryTest(server1, server2, "1001","server1");
 	}
-	
+
 	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException"})
-	@AllowedFFDC(value = { "java.util.concurrent.RejectedExecutionException", "com.ibm.ws.Transaction.JTA.HeuristicHazardException" })
-	// JDK8: Allow HeuristicHazardException
-	public void WSTXMPR014BFVT() throws Exception {
-		recoveryTest(server1, server2, "1402","server2");
+	@AllowedFFDC(value = {"javax.xml.ws.WebServiceException"})
+	// XAException should be expected here but halt is no longer halting the JVM so extra protocol
+	// messages can get through after the dumpState.
+	public void WSTXMPR010BFVT() throws Exception {
+		recoveryTest(server1, server2, "1002","server2");
 	}
-	
+
 	@Test
-	public void WSTXMPR014CFVT() throws Exception {
-		recoveryTest(server1, server2, "1403","both");
+	@ExpectedFFDC(value = {"javax.transaction.xa.XAException" })
+	// Can get HeuristicRollbackException if commit gets through to subordinate after resources were already dumped
+	@AllowedFFDC(value={"javax.transaction.HeuristicRollbackException"})
+	public void WSTXMPR010CFVT() throws Exception {
+		recoveryTest(server1, server2, "1003","both");
 	}
-	
-	@Test
-	public void WSTXMPR015AFVT() throws Exception {
-		recoveryTest(server1, server2, "1501","server1");
-	}
-	
-	@Test
-	public void WSTXMPR015BFVT() throws Exception {
-		recoveryTest(server1, server2, "1502","server2");
-	}
-	
-	@Test
-	public void WSTXMPR015CFVT() throws Exception {
-		recoveryTest(server1, server2, "1503","both");
-	}
-	
+
 	@Test
 	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
-	public void WSTXMPR016AFVT() throws Exception {
-		recoveryTest(server1, server2, "1601","server1");
+	public void WSTXMPR011AFVT() throws Exception {
+		recoveryTest(server1, server2, "1101","server1");
 	}
-	
+
 	@Test
 	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
-	public void WSTXMPR016BFVT() throws Exception {
-		recoveryTest(server1, server2, "1602","server2");
-	}
-	
-	@Test
-	@ExpectedFFDC(value = {"javax.transaction.xa.XAException", "javax.transaction.RollbackException"})
-	public void WSTXMPR016CFVT() throws Exception {
-		recoveryTest(server1, server2, "1603","both");
+	public void WSTXMPR011BFVT() throws Exception {
+		recoveryTest(server1, server2, "1102","server2");
 	}
 }
