+++ dev/com.ibm.ws.wsat.recovery_fat.multi.5/.gitignore
@@ -1,0 +1,1 @@
+/publish
