+++ dev/com.ibm.ws.wsat.recovery_fat.multi.5/build.gradle
@@ -1,0 +1,58 @@
+/*******************************************************************************
+ * Copyright (c) 2020, 2023 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ * 
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+
+// Define G:A:V coordinates of each dependency
+dependencies {
+  requiredLibs project(":com.ibm.ws.wsat.recovery_fat.lps"),
+               project(":com.ibm.ws.wsat.recovery_fat.multi.1"),
+               project(":com.ibm.ws.transaction.fat.util")
+}
+
+task copyCommonFiles {
+  doLast {
+    /*
+     * Copy common publish files.
+     */
+    copy {
+      from project(':com.ibm.ws.wsat.recovery_fat.lps').file('publish')
+      into new File('publish')
+    }
+  }
+}
+
+jar {
+  dependsOn copyCommonFiles
+}
+
+autoFVT.dependsOn copyCommonFiles
+
+clean.doLast {
+  if (file('publish').exists()){
+    file('publish').deleteDir()
+  }
+  if (file('build').exists()){
+    file('build').deleteDir()
+  }
+}
+
+autoFVT.doLast {
+  copy {
+    from project(':com.ibm.ws.wsat.recovery_fat.lps').autoFvtDir
+    into autoFvtDir
+    include 'test-applications/**'
+  }
+}
+
+autoFVT.dependsOn project(':com.ibm.ws.wsat.recovery_fat.lps').autoFVT
+
+addRequiredLibraries.dependsOn addJakartaTransformer
